重构插件整体结构, 区分安装器和utils
